requirements.txt
# Core dependencies for S&P 500 Financial Analysis
# Keep this minimal - only what we actually need

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Database
duckdb>=0.9.0

# Data validation
pydantic>=2.0.0

# Azure OpenAI integration
openai>=1.12.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0

